Refactor the text tokanizers into functions and introduce generic `Tokenizer`
